refactor(playwright): move _marketplace_service fixture to conftest
