+++ tests/playwright/test_marketplace_lifecycle.py
@@ -10,12 +10,11 @@ created on disk inside ``tmp_path`` by the ``_marketplace_service`` fixture.
 This guarantees a stable, predictable target with no dependency on a live
 marketplace server.
 
-The ``_marketplace_service`` fixture patches
+The ``_marketplace_service`` fixture (defined in ``conftest.py``) patches
 ``file_organizer.web.marketplace_routes._service`` so the in-process live
-server (started by the session-scoped ``live_server_url`` fixture in
-conftest.py) resolves to a ``MarketplaceService`` backed by a real but
-isolated local repo directory.  State written during a test (installed.json)
-lives inside ``tmp_path`` and is discarded when the fixture tears down.
+server resolves to a ``MarketplaceService`` backed by a real but isolated
+local repo directory.  State written during a test (installed.json) lives
+inside ``tmp_path`` and is discarded when the fixture tears down.
 
 Auth
 ----
@@ -32,15 +31,7 @@ Running
 
 from __future__ import annotations
 
-import json
-import zipfile
-from collections.abc import Iterator
-from pathlib import Path
-from unittest.mock import patch
-
 import pytest
-
-from file_organizer.plugins.marketplace import MarketplaceService, compute_sha256
 
 try:
     from playwright.sync_api import Page, expect
@@ -56,81 +47,6 @@ pytestmark = [
 ]
 
 _LOCATOR_TIMEOUT_MS = 10_000
-_STUB_PLUGIN_NAME = "fo-test-echo"
-_STUB_PLUGIN_VERSION = "1.0.0"
-
-_PLUGIN_PY = "\n".join(
-    [
-        "from file_organizer.plugins import Plugin, PluginMetadata",
-        "",
-        "class EchoPlugin(Plugin):",
-        "    def get_metadata(self):",
-        f"        return PluginMetadata(name='{_STUB_PLUGIN_NAME}', version='{_STUB_PLUGIN_VERSION}', author='tests', description='plugin')",
-        "    def on_load(self): pass",
-        "    def on_enable(self): pass",
-        "    def on_disable(self): pass",
-        "    def on_unload(self): pass",
-    ]
-)
-
-
-@pytest.fixture
-def _marketplace_service(tmp_path: Path) -> Iterator[str]:
-    """Seed a real local stub plugin repo and patch the marketplace _service() factory.
-
-    Creates a self-contained marketplace home directory inside ``tmp_path``:
-    - ``home/repository/`` contains ``fo-test-echo-1.0.0.zip`` and ``index.json``
-    - ``home/`` is the MarketplaceService home (installed.json lands here)
-
-    Patches ``file_organizer.web.marketplace_routes._service`` so that every
-    request handled by the in-process live server during the test receives a
-    real ``MarketplaceService`` backed by the seeded repo.
-
-    Yields:
-        The stub plugin name (``"fo-test-echo"``).
-    """
-    home = tmp_path / "home"
-    home.mkdir()
-    repo_dir = home / "repository"
-    repo_dir.mkdir()
-
-    archive_name = f"{_STUB_PLUGIN_NAME}-{_STUB_PLUGIN_VERSION}.zip"
-    archive_path = repo_dir / archive_name
-    with zipfile.ZipFile(archive_path, "w") as zf:
-        zf.writestr("plugin.py", _PLUGIN_PY)
-
-    metadata = {
-        "name": _STUB_PLUGIN_NAME,
-        "version": _STUB_PLUGIN_VERSION,
-        "author": "tests",
-        "description": f"{_STUB_PLUGIN_NAME} plugin",
-        "homepage": "https://example.invalid",
-        # Bare filename (no scheme): PluginRepository resolves it relative to
-        # _base_file_root, which is derived from repo_url=str(repo_dir).
-        "download_url": archive_name,
-        "checksum_sha256": compute_sha256(archive_path),
-        "size_bytes": archive_path.stat().st_size,
-        "dependencies": [],
-        "tags": ["utility"],
-        "category": "utility",
-        "license": "MIT",
-        "min_organizer_version": "2.0.0",
-        "max_organizer_version": None,
-        "downloads": 0,
-        "rating": 0.0,
-        "reviews_count": 0,
-    }
-    (repo_dir / "index.json").write_text(
-        json.dumps({"plugins": [metadata]}, indent=2), encoding="utf-8"
-    )
-
-    service = MarketplaceService(home_dir=home, repo_url=str(repo_dir))
-
-    with patch(
-        "file_organizer.web.marketplace_routes._service",
-        new=lambda: service,
-    ):
-        yield _STUB_PLUGIN_NAME
 
 
 # ---------------------------------------------------------------------------
